feat: validate login and redirect
