--- src/context/authGoogle.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const auth = createContext();
-
-export default auth;
